API.py
from urllib import response
from flask import Flask, request, session, jsonify
from utils.errors import APIError, handle_api_error
from flask_caching import Cache
from flask_sqlalchemy import SQLAlchemy
import logging
from logging.handlers import RotatingFileHandler
import os
from flask_session import Session
from flask_limiter import Limiter
from flask_limiter.util import get_remote_address
from flask_jwt_extended import JWTManager
from flask_cors import CORS
from flask_caching import Cache
import admin_api, adminlogin
import random
import json
import string
import urllib.parse
from flask_mysqldb import MySQL
from config import config
from utils.logger import setup_logger

from flask import Blueprint
bp = Blueprint('main', __name__)

# Initialize cache
cache = Cache(app, config={
    'CACHE_TYPE': 'simple',
    'CACHE_DEFAULT_TIMEOUT': 300
})

# Database connection pooling
app.config['SQLALCHEMY_DATABASE_URI'] = f"mysql://{app.config['MYSQL_USER']}:{app.config['MYSQL_PASSWORD']}@{app.config['MYSQL_HOST']}/{app.config['MYSQL_DB']}"
app.config['SQLALCHEMY_POOL_SIZE'] = 10
app.config['SQLALCHEMY_POOL_TIMEOUT'] = 30
db = SQLAlchemy(app)

# Setup logging
if not os.path.exists('logs'):
    os.makedirs('logs')
file_handler = RotatingFileHandler('logs/app.log', maxBytes=10240, backupCount=10)
file_handler.setFormatter(logging.Formatter(
    '%(asctime)s %(levelname)s: %(message)s [in %(pathname)s:%(lineno)d]'
))
file_handler.setLevel(logging.INFO)
app.logger.addHandler(file_handler)
app.logger.setLevel(logging.INFO)
app.logger.info('Flask Backend Framework startup')

# Register error handlers
app.register_error_handler(APIError, handle_api_error)
app.config.from_object(config['development'])

# Initialize extensions
mysql = MySQL(app)
jwt = JWTManager(app)
cache = Cache(app, config={'CACHE_TYPE': 'simple'})
CORS(app)
limiter = Limiter(
    app=app,
    key_func=get_remote_address,
    default_limits=["200 per day", "50 per hour"]
)

# Setup session
Session(app)

# Setup logging
setup_logger(app)

# Error handlers
@app.errorhandler(404)
def not_found_error(error):
    app.logger.error(f'Page not found: {request.url}')
    return jsonify({"success": False, "message": "Resource not found"}), 404

@app.errorhandler(500)
def internal_error(error):
    app.logger.error(f'Server Error: {error}')
    return jsonify({"success": False, "message": "Internal server error"}), 500

# Rate limit decorators for sensitive endpoints
@limiter.limit("5 per minute")
@app.route('/generate_otp', methods=['POST'])
def trigger_generate_otp():
    try:
        if not request.form.get('phone'):
            raise APIError("Phone number is required", 400)
        response = otp.generate_otp()
        app.logger.info(f"OTP generated for phone: {request.form.get('phone')}")
        return response
    except Exception as e:
        app.logger.error(f"OTP generation failed: {str(e)}")
        raise APIError("Failed to generate OTP", 500)

[EXISTING_CODE_PRESERVED]

if __name__ == "__main__":
    from app import create_app
    app = create_app()
    app.run(debug=True)