requirements.txt
Flask==2.0.1
Flask-Session==0.4.0
Flask-MySQL==1.5.2
boto3==1.26.137
itsdangerous==2.0.1
python-dotenv==0.19.0
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Caching==2.1.0
Flask-SQLAlchemy==3.1.1
marshmallow==3.20.1
pytest==7.4.3
